Replaced make swig-all with make
