+++ tests/data/scenario_baseline_parameters.csv
@@ -1,0 +1,2 @@
+rng_seed, param_id, n_total, mean_work_interactions_child, mean_work_interactions_adult, mean_work_interactions_elderly, daily_fraction_work, mean_random_interactions_child, sd_random_interactions_child, mean_random_interactions_adult, sd_random_interactions_adult, mean_random_interactions_elderly, sd_random_interactions_elderly, random_interaction_distribution, child_network_adults, elderly_network_adults, days_of_interactions, end_time, n_seed_infection, mean_infectious_period, sd_infectious_period, infectious_rate, mean_time_to_symptoms, sd_time_to_symptoms, mean_time_to_hospital, mean_time_to_critical, mean_time_to_recover, sd_time_to_recover, mean_time_to_death, sd_time_to_death, fraction_asymptomatic_0_9, fraction_asymptomatic_10_19, fraction_asymptomatic_20_29, fraction_asymptomatic_30_39, fraction_asymptomatic_40_49, fraction_asymptomatic_50_59, fraction_asymptomatic_60_69, fraction_asymptomatic_70_79, fraction_asymptomatic_80, asymptomatic_infectious_factor, mild_fraction_0_9, mild_fraction_10_19, mild_fraction_20_29, mild_fraction_30_39, mild_fraction_40_49, mild_fraction_50_59, mild_fraction_60_69, mild_fraction_70_79, mild_fraction_80, mild_infectious_factor, mean_asymptomatic_to_recovery, sd_asymptomatic_to_recovery, household_size_1, household_size_2, household_size_3, household_size_4, household_size_5, household_size_6, population_0_9, population_10_19, population_20_29, population_30_39, population_40_49, population_50_59, population_60_69, population_70_79, population_80, daily_non_cov_symptoms_rate, relative_susceptibility_0_9, relative_susceptibility_10_19, relative_susceptibility_20_29, relative_susceptibility_30_39, relative_susceptibility_40_49, relative_susceptibility_50_59, relative_susceptibility_60_69, relative_susceptibility_70_79, relative_susceptibility_80, relative_transmission_household, relative_transmission_workplace, relative_transmission_random, hospitalised_fraction_0_9, hospitalised_fraction_10_19, hospitalised_fraction_20_29, hospitalised_fraction_30_39, hospitalised_fraction_40_49, hospitalised_fraction_50_59, hospitalised_fraction_60_69, hospitalised_fraction_70_79, hospitalised_fraction_80, critical_fraction_0_9, critical_fraction_10_19, critical_fraction_20_29, critical_fraction_30_39, critical_fraction_40_49, critical_fraction_50_59, critical_fraction_60_69, critical_fraction_70_79, critical_fraction_80, fatality_fraction_0_9, fatality_fraction_10_19, fatality_fraction_20_29, fatality_fraction_30_39, fatality_fraction_40_49, fatality_fraction_50_59, fatality_fraction_60_69, fatality_fraction_70_79, fatality_fraction_80, mean_time_hospitalised_recovery, sd_time_hospitalised_recovery, mean_time_critical_survive, sd_time_critical_survive, icu_allocation_0_9, icu_allocation_10_19, icu_allocation_20_29, icu_allocation_30_39, icu_allocation_40_49, icu_allocation_50_59, icu_allocation_60_69, icu_allocation_70_79, icu_allocation_80, quarantine_length_self, quarantine_length_traced, quarantine_length_positive, quarantine_dropout_self, quarantine_dropout_traced, quarantine_dropout_positive, test_on_symptoms, test_on_traced, trace_on_symptoms, trace_on_positive, retrace_on_positive, quarantine_on_traced, traceable_interaction_fraction, tracing_network_depth, allow_clinical_diagnosis, quarantine_household_on_positive, quarantine_household_on_symptoms, quarantine_household_on_traced_positive, quarantine_household_on_traced_symptoms, quarantine_household_contacts_on_positive, quarantine_household_contacts_on_symptoms, quarantined_daily_interactions, quarantine_days, quarantine_smart_release_day, hospitalised_daily_interactions, test_insensitive_period, test_order_wait, test_result_wait, self_quarantine_fraction, app_users_fraction_0_9, app_users_fraction_10_19, app_users_fraction_20_29, app_users_fraction_30_39, app_users_fraction_40_49, app_users_fraction_50_59, app_users_fraction_60_69, app_users_fraction_70_79, app_users_fraction_80, app_turn_on_time, lockdown_occupation_multiplier_primary_network, lockdown_occupation_multiplier_secondary_network, lockdown_occupation_multiplier_working_network, lockdown_occupation_multiplier_retired_network, lockdown_occupation_multiplier_elderly_network, lockdown_random_network_multiplier, lockdown_house_interaction_multiplier, lockdown_time_on, lockdown_time_off, lockdown_elderly_time_on, lockdown_elderly_time_off, testing_symptoms_time_on, testing_symptoms_time_off, intervention_start_time, hospital_on
+1, 1, 20000, 10, 7, 3, 0.5, 2, 2, 4, 4, 3, 3, 1, 0.2, 0.2, 10, 200, 5, 6, 2.5, 5.75, 6, 2.5, 5.14, 2.5, 12, 5, 6, 2, 0.18, 0.18, 0.18, 0.18, 0.18, 0.18, 0.18, 0.18, 0.18, 0.29, 0.79, 0.79, 0.73, 0.68, 0.65, 0.59, 0.53, 0.41, 0.27, 0.48, 15, 5, 7452, 9936, 4416, 4140, 1104, 552, 8054000, 7528000, 8712000, 8835000, 8500000, 8968000, 7069000, 5488000, 3281000, 0.002, 0.71, 0.74, 0.79, 0.87, 0.98, 1.11, 1.26, 1.45, 1.66, 2, 1, 1, 0.01, 0.0267, 0.0367, 0.0633, 0.08, 0.12, 0.1533, 0.16, 0.1367, 0.05, 0.05, 0.05, 0.05, 0.063, 0.122, 0.274, 0.432, 0.709, 0.33, 0.25, 0.5, 0.5, 0.5, 0.69, 0.65, 0.88, 1, 8, 3, 4, 2, 1, 1, 1, 1, 1, 1, 1, 1, 0.5, 7, 14, 14, 0.02, 0.02, 0.02, 0, 0, 0, 0, 0, 0, 0.8, 0, 1, 0, 0, 0, 0, 0, 0, 0, 7, 0, 0, 3, 1, 1, 0, 0.0, 0.9, 0.96, 0.95, 0.91, 0.81, 0.64, 0.41, 0.27, 10000, 0.2, 0.2, 0.2, 0.2, 0.2, 0.2, 1.5, 10000, 10000, 10000, 10000, 10000, 10000, 0, 1
