+++ tests/data/trace_tokens_Run1.csv
@@ -1,0 +1,1 @@
+time,days_since_index,index_ID,index_status,days_since_contact,traced_ID,traced_status,traced_infector_ID,traced_time_infected
